// src/app/Models/User.ts
export interface User{
    id: string;
    username: string;
    displayName: string;
    privateProfile: boolean;
    userImage: string;
    followers: Number;
    following: Number;
    noOfPosts: Number;
    bio: string;
    website: string;
    email: string;
    phoneNumber: string;
    gender: string;
}